Add Generator(Device device) in C++ API
[ghstack-poisoned]
