+++ aten/src/ATen/CPUGenerator.cpp
@@ -7,26 +7,6 @@ namespace at {
 namespace detail {
 
 /**
- * PyTorch maintains a collection of default generators that get
- * initialized once. The purpose of these default generators is to
- * maintain a global running state of the pseudo random number generation,
- * when a user does not explicitly mention any generator.
- * getDefaultCPUGenerator gets the default generator for a particular
- * device.
- */
-const Generator& getDefaultCPUGenerator() {
-  static auto default_gen_cpu = createCPUGenerator(c10::detail::getNonDeterministicRandom());
-  return default_gen_cpu;
-}
-
-/**
- * Utility to create a CPUGenerator. Returns a shared_ptr
- */
-Generator createCPUGenerator(uint64_t seed_val) {
-  return make_generator<CPUGenerator>(seed_val);
-}
-
-/**
  * Helper function to concatenate two 32 bit unsigned int
  * and return them as a 64 bit unsigned int
  */
