+++ torch/csrc/Generator.cpp
@@ -51,21 +51,7 @@ static PyObject * THPGenerator_pynew(PyTypeObject *type, PyObject *args, PyObjec
   auto device = r.deviceWithDefault(0, at::Device(at::kCPU));
 
   THPGeneratorPtr self((THPGenerator *)type->tp_alloc(type, 0));
-#ifdef USE_CUDA
-  if (device.type() == at::kCPU) {
-    self->cdata = make_generator<CPUGenerator>();
-  } else if (device.type() == at::kCUDA){
-    self->cdata = make_generator<CUDAGenerator>(device.index());
-  } else {
-    AT_ERROR("Device type ", c10::DeviceTypeName(device.type()),
-             " is not supported for torch.Generator() api.");
-  }
-#else
-  TORCH_CHECK(device.type() == at::kCPU,
-              "Device type ", c10::DeviceTypeName(device.type()),
-              " is not supported for torch.Generator() api.");
-  self->cdata = make_generator<CPUGenerator>();
-#endif
+  self->cdata = Generator(device);
   return (PyObject*)self.release();
   END_HANDLE_TH_ERRORS
 }
